Created the explainer Notebook.
The Notebook should contain all the behind the scenes stuff, details on the dataset, why you’ve selected this particular dataset, explanations of your choices regarding network analysis, etc. You should link to the notebook from the site. The idea is that you can create much more complex, fun and interactive analysis (and visualizations) on line. So the website is a way for you to present your work in a way that everyone can understand it … including dynamic visualizations, interactive analysis, etc, etc … that would not work on a piece of paper.